refactor(routes): inline remaining thin wrappers, adopt suspense queries
Inline StageManagement and Timeline into their route files, and switch
FestivalDetail, admins, artists, and EditionSelection to the loader
ensureQueryData + useSuspenseQuery idiom, removing dead manual isLoading
branches now that data is guaranteed cached by the loader/beforeLoad.
