--- src/pages/EditionView/tabs/ScheduleTab/horizontal/Timeline.tsx
@@ -1,116 +1,0 @@
-import { useMemo } from "react";
-import { useSuspenseQuery } from "@tanstack/react-query";
-import { useRouteContext } from "@tanstack/react-router";
-import { useScheduleData } from "@/hooks/useScheduleData";
-import {
-  calculateScheduleWindow,
-  calculateTimelineData,
-} from "@/lib/timelineCalculator";
-import { TimelineContainer } from "./TimelineContainer";
-import { useFestivalEdition } from "@/contexts/FestivalEditionContext";
-import { useSetsByEditionQuery as useEditionSetsQuery } from "@/api/sets/useSetsByEdition";
-import { useTimelineUrlState } from "@/hooks/useTimelineUrlState";
-import { useNow } from "@/hooks/useNow";
-import { filterScheduleDays } from "@/lib/scheduleFilter";
-import { stagesByEditionQuery } from "@/api/stages/useStagesByEdition";
-import { useScheduleReveal } from "@/hooks/useScheduleReveal";
-import { ScheduleNotRevealedPlaceholder } from "../ScheduleNotRevealedPlaceholder";
-import { useAuth } from "@/contexts/AuthContext";
-import { useUserVotesQuery } from "@/api/voting/useUserVotesQuery";
-
-export function Timeline() {
-  const { festival } = useFestivalEdition();
-  const { edition } = useRouteContext({
-    from: "/festivals/$festivalSlug/editions/$editionSlug/schedule/timeline",
-  });
-  const { canShowTime } = useScheduleReveal();
-  const now = useNow();
-  const { data: editionSets = [], isLoading: setsLoading } =
-    useEditionSetsQuery(edition.id);
-  const { data: stages } = useSuspenseQuery(stagesByEditionQuery(edition.id));
-  const { user } = useAuth();
-  const { data: userVotes } = useUserVotesQuery(user?.id);
-
-  const { scheduleDays } = useScheduleData({
-    sets: editionSets,
-    stages,
-    timezone: festival.timezone,
-  });
-  const {
-    day: selectedDay,
-    time: selectedTime,
-    stages: selectedStages,
-    votes: selectedVotes,
-  } = useTimelineUrlState("timeline");
-
-  const scheduleWindow = useMemo(
-    () => calculateScheduleWindow(scheduleDays),
-    [scheduleDays],
-  );
-
-  const timelineData = useMemo(() => {
-    if (!edition.start_date || !edition.end_date) {
-      return null;
-    }
-
-    const filteredScheduleDays = filterScheduleDays(
-      scheduleDays,
-      {
-        day: selectedDay,
-        time: selectedTime,
-        stages: selectedStages,
-        voteTypes: selectedVotes,
-        userVotes,
-      },
-      festival.timezone,
-    );
-
-    return calculateTimelineData(
-      new Date(edition.start_date),
-      new Date(edition.end_date),
-      filteredScheduleDays,
-      stages,
-    );
-  }, [
-    edition,
-    scheduleDays,
-    selectedDay,
-    selectedTime,
-    selectedStages,
-    selectedVotes,
-    userVotes,
-    stages,
-    festival.timezone,
-  ]);
-
-  if (setsLoading) {
-    return (
-      <div className="text-center text-purple-300 py-12">
-        <p>Loading horizontal timeline...</p>
-      </div>
-    );
-  }
-
-  if (!timelineData) {
-    return (
-      <div className="text-center text-purple-300 py-12">
-        <p>Festival dates not available yet.</p>
-      </div>
-    );
-  }
-
-  if (!canShowTime) {
-    return <ScheduleNotRevealedPlaceholder />;
-  }
-
-  return (
-    <TimelineContainer
-      timelineData={timelineData}
-      timezone={festival.timezone}
-      scheduleDays={scheduleDays}
-      selectedDay={selectedDay}
-      scheduleWindow={scheduleWindow}
-      now={now}
-    />
-  );
-}
